Make the configuration of the version more user-friendly
The client expects the version to be a string '3.0'. This change allows
to configure the version as a float by formatting it properly, to make
it possible to use the unquoted syntax in Yaml, which will produce a
float.
Note that the config file in the test fixtures was already using a
float, but the tests were not failing because the service was not
instantiated.